# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hotel/book/initiate": {
      "post": {
        "description": "Retrieves available room types, rates, and pricing information for a specific hotel based on check-in/check-out dates and guest requirements.",
        "operationId": "HotelBookController_initiate",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelBookInitiateDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Initiate hotel book",
        "tags": [
          "Hotel"
        ]
      }
    },
    "/hotel/book/confirmation": {
      "post": {
        "description": "Retrieves available room types, rates, and pricing information for a specific hotel based on check-in/check-out dates and guest requirements.",
        "operationId": "HotelBookController_bookConfirmation",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelBookConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Hotel book confirmation",
        "tags": [
          "Hotel"
        ]
      }
    },
...
    "/flight/order-detail": {
      "post": {
        "operationId": "OrderDetailController_getOrderDetails",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDetailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get order details",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/book/initiate": {
      "post": {
        "operationId": "BookController_bookingInitiate",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Book a flight",
        "tags": [
          "Flight"
        ]
      }
    },
    "/flight/book/confirmation": {
      "post": {
        "operationId": "BookController_bookingConfirmation",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Book a flight",
        "tags": [
          "Flight"
        ]
      }
    },
...
  },
  "info": {
    "title": "API SERVICE",
    "description": "## Travel API Documentation\n\nWelcome to the **API SERVICE** documentation. This guide provides comprehensive guidelines and detailed instructions on how to effectively integrate and utilize the Travel API.\n\n### Overview\nThis documentation includes various sections covering:\n- **Authentication Methods**: Learn how to authenticate using Bearer and Cookie authentication.\n- **Endpoint Descriptions**: Detailed information on each API endpoint.\n- **Usage Examples**: Practical examples to help you get started quickly.\n- **Best Practices**: Tips and recommendations for efficient implementation.\n\n### Purpose\nWhether you are a developer looking to incorporate travel-related functionalities or a business aiming to enhance your travel services, this documentation serves as a valuable resource to guide you through the process.\n\n### Getting Started\nTo get started, ensure you have the necessary authentication tokens and follow the endpoint descriptions to make your first API call.\n\nFor any questions or support, please refer to the contact section at the end of this documentation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
    "schemas": {
...
      "PassengerDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Passenger type (adult, child, infant)",
            "example": "adult"
          },
          "title": {
            "type": "string",
            "description": "Passenger title (Mr, Miss, Mrs, Ms)",
            "example": "Mr"
          },
          "roomId": {
            "type": "number",
            "description": "Passenger roomId (1,2,3)",
            "example": 1
          },
          "age": {
            "type": "number",
            "description": "Passenger age (required for child/infant)",
            "example": 25
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "dialCode": {
            "type": "string",
            "description": "dial Code",
            "example": "+91"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "+1234567890"
          },
          "nationality": {
            "type": "string",
            "description": "Nationality code",
            "example": "US"
          }
        },
        "required": [
          "type",
          "title",
          "roomId",
          "firstName",
          "lastName"
        ]
      },
      "PaymentDetailsDto": {
        "type": "object",
        "properties": {
          "gatewayName": {
            "type": "string",
            "description": "Payment gateway name",
            "example": "stripe"
          },
          "paymentType": {
            "type": "string",
            "description": "Payment type",
            "example": "credit_card"
          },
          "totalAmount": {
            "type": "number",
            "description": "Amount",
            "example": 360
          },
          "cashAmount": {
            "type": "number",
            "description": "Cash amount",
            "example": 360
          },
          "priceHashKey": {
            "type": "string",
            "description": "Price hash key",
            "example": "price_hash_key"
          },
          "paymentToken": {
            "type": "string",
            "description": "Payment token",
            "example": "tok_1234567890abcdef"
          },
          "paymentLogId": {
            "type": "string",
            "description": "Payment log ID",
            "example": "log_1234567890abcdef"
          }
        },
        "required": [
          "gatewayName",
          "paymentType",
          "totalAmount",
          "cashAmount",
          "priceHashKey",
          "paymentToken",
          "paymentLogId"
        ]
      },
      "ContactDetailsDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "User title Mr, Miss, and Mrs",
            "example": "Mr"
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "gender": {
            "type": "string",
            "description": "Contact person/user gender male, female and other",
            "example": "male"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "dialCode": {
            "type": "string",
            "description": "Dialer Code as country wise",
            "example": "+91"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "+1234567890"
          },
          "addressLine1": {
            "type": "string",
            "description": "Address line 1",
            "example": "123 Main Street"
          },
          "addressLine2": {
            "type": "string",
            "description": "Address line 2",
            "example": "Apt 4B"
          },
          "city": {
            "type": "string",
            "description": "City",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "description": "State",
            "example": "NY"
          },
          "country": {
            "type": "string",
            "description": "Country",
            "example": "United States"
          },
          "postalCode": {
            "type": "string",
            "description": "Postal code",
            "example": "10001"
          }
        },
        "required": [
          "title",
          "firstName",
          "lastName",
          "gender",
          "email",
          "dialCode",
          "phone",
          "addressLine1",
          "city",
          "state",
          "country",
          "postalCode"
        ]
      },
      "HotelBookInitiateDto": {
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "string",
            "description": "Hotel ID",
            "example": "107953"
          },
          "supplierCode": {
            "type": "string",
            "description": "Supplier code",
            "example": "TBO"
          },
          "searchReqId": {
            "type": "string",
            "description": "Search request ID from previous search",
            "example": "search_req_12345"
          },
          "rateKey": {
            "type": "string",
            "description": "Unique booking identifier",
            "example": "booking_12345"
          },
          "passengers": {
            "description": "Passenger details",
            "example": [
              {
                "type": "adult",
                "roomId": 1,
                "firstName": "John",
                "lastName": "Doe",
                "email": "john.doe@example.com",
                "phone": "+1234567890",
                "nationality": "US"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PassengerDto"
            }
          },
          "paymentDetails": {
            "description": "Payment details",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentDetailsDto"
              }
            ]
          },
          "contactDetails": {
            "description": "Contact details",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactDetailsDto"
              }
            ]
          }
        },
        "required": [
          "hotelId",
          "supplierCode",
          "searchReqId",
          "rateKey",
          "passengers",
          "paymentDetails",
          "contactDetails"
        ]
      },
      "HotelBookConfirmationDto": {
        "type": "object",
        "properties": {
          "bookingRefId": {
            "type": "string",
            "description": "Supplier booking ID",
            "example": "HB-2343246"
          },
          "searchReqId": {
            "type": "string",
            "description": "Search Request ID",
            "example": "1234567890"
          },
          "paymentLogId": {
            "type": "string",
            "description": "Payment log ID",
            "example": "1234567890"
          }
        },
        "required": [
          "bookingRefId",
          "searchReqId",
          "paymentLogId"
        ]
      },
      "SearchAirLegs": {
        "type": "object",
        "properties": {
          "departureDate": {
            "type": "string",
            "description": "Departure date for the flight",
            "example": "2025-06-15",
            "format": "date"
          },
          "origin": {
            "type": "string",
            "description": "Origin airport code",
            "example": "NYC"
          },
          "destination": {
            "type": "string",
            "description": "Destination airport code",
            "example": "LAX"
          }
        },
        "required": [
          "departureDate",
          "origin",
          "destination"
        ]
      },
...
      "OrderDetailDto": {
        "type": "object",
        "properties": {
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          },
          "bookingDetails": {
            "description": "bookingDetails will be here.",
            "example": [
              {
                "orderStatus": 1,
                "pnr": "abcdf",
                "orderNo": "12345",
                "firstName": "test",
                "lastName": "user"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "searchReqID": {
            "type": "string",
            "description": "Search request ID",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "searchAirLegs": {
            "description": "Search air legs information",
            "example": [
              {
                "origin": "NYC",
                "destination": "LAX"
              },
              {
                "origin": "LAX",
                "destination": "NYC"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchAirLegs"
            }
          },
          "mode": {
            "type": "string",
            "description": "Provider mode",
            "example": "Test",
            "enum": [
              "Production",
              "Test",
              "Sandbox"
            ]
          }
        },
        "required": [
          "providerCode",
          "bookingDetails",
          "searchReqID",
          "searchAirLegs",
          "mode"
        ]
      },
      "PassengerDetail": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Passenger title",
            "example": "Mr",
            "enum": [
              "Mr",
              "Mrs",
              "Ms",
              "Dr",
              "Prof"
            ]
          },
          "firstName": {
            "type": "string",
            "description": "Passenger first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Passenger last name",
            "example": "Doe"
          }
        },
        "required": [
          "title",
          "firstName",
          "lastName"
        ]
      },
...
      "Passenger": {
        "type": "object",
        "properties": {
          "passengerType": {
            "type": "string",
            "description": "Type of passenger",
            "example": "ADT",
            "enum": [
              "ADT",
              "CHD",
              "INF"
            ]
          },
          "gender": {
            "type": "string",
            "description": "Passenger gender",
            "example": "M",
            "enum": [
              "M",
              "F"
            ]
          },
          "passengerDetail": {
            "description": "Passenger personal details",
            "example": {
              "title": "Mr",
              "firstName": "John",
              "lastName": "Doe"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/PassengerDetail"
              }
            ]
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Passenger date of birth",
            "example": "1990-01-15",
            "format": "date"
          },
          "document": {
            "description": "Travel document information",
            "example": {
              "documentType": "Passport",
              "documentNumber": "A1234567",
              "expiryDate": "2030-12-31",
              "country": "US"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/Document"
              }
            ]
          },
          "nationality": {
            "type": "string",
            "description": "Passenger nationality",
            "example": "US"
          },
          "mobile": {
            "type": "string",
            "description": "Passenger mobile number",
            "example": "1234567890"
          },
          "mobileCountryCode": {
            "type": "string",
            "description": "Mobile country code",
            "example": "+1"
          }
        },
        "required": [
          "passengerType",
          "gender",
          "passengerDetail",
          "dateOfBirth",
          "nationality",
          "mobile",
          "mobileCountryCode"
        ]
      },
...
      "BookDto": {
        "type": "object",
        "properties": {
          "solutionId": {
            "type": "string",
            "description": "Solution ID for booking",
            "example": "7e3294817ad26fe6a885ca3b57312dbd"
          },
          "searchReqID": {
            "type": "string",
            "description": "Search request ID",
            "example": "efc080e1-3989-427d-abd4-edc7c04d27ab",
            "format": "uuid"
          },
          "trackingId": {
            "type": "string",
            "description": "Tracking ID",
            "example": "bdd3ff73-5ee7-49fc-bafa-34ba0666bfb7"
          },
          "providerCode": {
            "type": "string",
            "description": "Provider code",
            "example": "PK"
          },
          "fareType": {
            "type": "string",
            "description": "Fare type",
            "example": "PUBLIC",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "CORPORATE"
            ]
          },
          "airTripType": {
            "type": "string",
            "description": "airTripType needed here.",
            "example": "oneway"
          },
          "airlineType": {
            "type": "string",
            "description": "airlineType received in search result, could be LCC, Non-LCC GDS etc.",
            "example": "LCC"
          },
          "passengers": {
            "description": "Passenger information",
            "example": [
              {
                "passengerType": "ADT",
                "gender": "M",
                "passengerDetail": {
                  "title": "Mr",
                  "firstName": "John",
                  "lastName": "Doe"
                },
                "dateOfBirth": "1990-01-15",
                "document": {
                  "documentType": "P",
                  "documentNumber": "A1234567",
                  "expiryDate": "2030-12-31",
                  "country": "US"
                },
                "nationality": "US",
                "mobile": "1234567890",
                "mobileCountryCode": "+1"
              },
              {
                "passengerType": "ADT",
                "gender": "F",
                "passengerDetail": {
                  "title": "Ms",
                  "firstName": "Jane",
                  "lastName": "Doe"
                },
                "dateOfBirth": "1992-03-20",
                "document": {
                  "documentType": "P",
                  "documentNumber": "B7654321",
                  "expiryDate": "2030-12-31",
                  "country": "US"
                },
                "nationality": "US",
                "mobile": "1234567891",
                "mobileCountryCode": "+1"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Passenger"
            }
          },
          "contact": {
            "description": "Contact information",
            "example": {
              "title": "Mr",
              "firstName": "John",
              "lastName": "Doe",
              "gender": "M",
              "email": "john.doe@example.com",
              "mobile": "1234567890",
              "mobileCountryCode": "+1",
              "postalCode": "10001"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/ContactInfo"
              }
            ]
          },
          "routes": {
            "description": "Route information",
            "example": [
              {
                "route_0": [
                  {
                    "airline": "AA",
                    "departureCode": "LGA",
                    "departureDate": "2026-01-15",
                    "departureTime": "09:29 AM",
                    "arrivalCode": "ORD",
                    "arrivalDate": "2026-01-15",
                    "arrivalTime": "11:10 AM",
                    "bookingCode": "Q",
                    "flightNum": "3234"
                  },
                  {
                    "airline": "AA",
                    "departureCode": "ORD",
                    "departureDate": "2026-01-15",
                    "departureTime": "02:14 PM",
                    "arrivalCode": "LAX",
                    "arrivalDate": "2026-01-15",
                    "arrivalTime": "04:52 PM",
                    "bookingCode": "Q",
                    "flightNum": "1204"
                  }
                ]
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Routes"
            }
          }
        },
        "required": [
          "solutionId",
          "searchReqID",
          "trackingId",
          "providerCode",
          "fareType",
          "airTripType",
          "airlineType",
          "passengers",
          "contact",
          "routes"
        ]
      },
      "BookConfirmationDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "description": "Booking ID",
            "example": "1234567890"
          },
          "bookingLogId": {
            "type": "string",
            "description": "Booking log ID",
            "example": "1234567890"
          }
        },
        "required": [
          "bookingId",
          "bookingLogId"
        ]
      },
...
    }
  }
}